+++ .prettierrc.cjs
@@ -1,5 +1,5 @@
 /** @type {import("@types/prettier").Options} */
 
 module.exports = {
-  ...require('@julian_cataldo/astro-configs/prettier-astro.cjs'),
+  ...require('webdev-configs/prettier-astro.cjs'),
 };