+++ app/BaseDocument/Props.ts
@@ -2,6 +2,11 @@
 
 export interface Props extends astroHTML.JSX.HTMLAttributes {
   /**
+   * **Default**: `'en'`
+   * */
+  lang?: string;
+
+  /**
    * Provides early JS detection for further CSS use
    * by adding `js` class critically to HTML element.
    *