+++ components/CodeEditor/YAMLEditor.Props.ts
@@ -1,0 +1,10 @@
+export interface Props extends astroHTML.JSX.HTMLAttributes {
+  value?: string;
+
+  extractMarkdownFrontmatter?: boolean;
+
+  schemaUrlRef?: string;
+
+  // TODO:
+  // schemaObject?: string;
+}