+++ configs/eslint-ts.cjs
@@ -57,6 +57,8 @@ module.exports = {
         'react/jsx-filename-extension': 'off',
 
         'import/prefer-default-export': 'off',
+
+        'import/extensions': 'off',
         // 'import/extensions': [
         //   'error',
         //   'ignorePackages',