+++ components/TerminalPlayer/CHANGELOG.md
@@ -3,18 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.1.1](https://github.com/JulianCataldo/web-garden/compare/astro-terminal-player@0.1.0...astro-terminal-player@0.1.1) (2022-10-17)
+## [0.1.5](https://github.com/JulianCataldo/web-garden/compare/astro-terminal-player@0.1.4...astro-terminal-player@0.1.5) (2022-11-19)
 
+**Note:** Version bump only for package astro-terminal-player
+
+## [0.1.4](https://github.com/JulianCataldo/web-garden/compare/astro-terminal-player@0.1.3...astro-terminal-player@0.1.4) (2022-10-21)
+
+**Note:** Version bump only for package astro-terminal-player
+
+## [0.1.3](https://github.com/JulianCataldo/web-garden/compare/astro-terminal-player@0.1.2...astro-terminal-player@0.1.3) (2022-10-21)
+
+**Note:** Version bump only for package astro-terminal-player
+
+## [0.1.2](https://github.com/JulianCataldo/web-garden/compare/astro-terminal-player@0.1.1...astro-terminal-player@0.1.2) (2022-10-18)
 
 ### Bug Fixes
 
-* remove package prefix, simplify import method ([73bd31b](https://github.com/JulianCataldo/web-garden/commit/73bd31bf1f501624036a74a3f19c5bf83cc9c0a4))
+- note + vite config for client side js error ([3a5c6d5](https://github.com/JulianCataldo/web-garden/commit/3a5c6d52a58623b49190f4988774024a4802b12a))
 
+## [0.1.1](https://github.com/JulianCataldo/web-garden/compare/astro-terminal-player@0.1.0...astro-terminal-player@0.1.1) (2022-10-17)
 
+### Bug Fixes
+
+- remove package prefix, simplify import method ([73bd31b](https://github.com/JulianCataldo/web-garden/commit/73bd31bf1f501624036a74a3f19c5bf83cc9c0a4))
 
 ## 0.1.0 (2022-10-15)
 
-
 ### Features
 
-* init terminal player component ([886e0fe](https://github.com/JulianCataldo/web-garden/commit/886e0feef9481b8a010576de1faa205ba6f5fd57))
+- init terminal player component ([886e0fe](https://github.com/JulianCataldo/web-garden/commit/886e0feef9481b8a010576de1faa205ba6f5fd57))